The size of the detect points were enlarged, and this fact may had chance the quality of the matches. The minimum distance between points were enlarged too, and this improved a lot the velocity to detect them.
Were created a change history manager (TrackPointsManager.h), for recover the lost tracked points. The matching recover resulting is not implemented yet, but the matches appear be good.

In next interaction should be removed the detected key points, that have a coincident track points, to minimize the error chance of wrong matches when trying recuperate the lost tracking points.

The metrics of matching recover points should be record, adding some interference of the user in the choice of the best matches, to verify what's is the best formula to get the best recover track points from matchs.

It's possible that will be create a verification method for evaluation of each track point history features, for be sure that them are not wrong marked as true.